requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
idna==3.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
python-dotenv==0.19.0
PyYAML==5.4.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.26.0
urllib3==1.26.6
